pyproject.toml
[project]
name = "socials-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "streamlit>=1.56.0",
    "openai>=1.0.0",
    "pillow>=10.0.0",
    "requests>=2.0.0",
]
